Use double to store energy in hits (#193)
* Use double to store energy in RMGGermaniumDetector, RMGScintillatorDetector, RMGOpticalDetector

* prevent this problem from happening again (add compiler warning for this)

* fix compiler warnings in RMGGeneratorCosmicMuons

+++ include/RMGOpticalDetector.hh
@@ -47,7 +47,7 @@ class RMGOpticalDetectorHit : public G4VHit {
 
     G4TouchableHandle detector_touchable; // for Draw()
     int detector_uid = -1;
-    float photon_wavelength = 0.;
+    double photon_wavelength = 0.;
     double global_time = -1;
 };
